Update start of debugging session mechanism
Before, a target remote command would start the guest VM
and would let it wait at the entry point. This commit
offers the user the responsibility of issuing a continue
command in the beginning so that the VM starts and executes
until reaching the entry point. This is also aligned
with the practice used in IDEs of starting the debugging
process with a default continue.

Signed-off-by: Marius-Cristian Baciu <baciumar@amazon.com>